% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_fieldbook.R
\name{add_cl}
\alias{add_cl}
\title{Creation of field design using factor or treatments}
\usage{
add_cl(fb, design_abr = "CRD", factor_lvl)
}
\arguments{
\item{fb}{field book design}

\item{design_abr}{design abbreviation}

\item{factor_lvl}{vector of factors or treatments.}
}
\description{
User may create field designs using different levels of a factor or treatment. For example if the factor has two levels;
lets say, irrigation and non-irrigation, two field books will be created one with irrigation and the another one with non-irrigations.
Only available for unreplicated designs, complete randomized designs `(CRD)`, completely block desing `(RCBD)`,
latin square designs `(LSD)`, Alpha design `(AD)`, Wescott design `(WD)` and Augmented Block Designs `(ABD)`
All those levels will be attached in a column named FACTOR.
}
\author{
Omar Benites
}
