% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{add_fieldbook_sheet}
\alias{add_fieldbook_sheet}
\title{Add values to Fieldbook Sheet Data}
\usage{
add_fieldbook_sheet(file, fieldbook)
}
\arguments{
\item{file}{The ID of fieldbook excel}

\item{fieldbook}{A data.frame with the fieldbook data}
}
\description{
This function allows to paste values into Fieldbook Excel Sheet
}

