% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_fieldbook.R
\name{add_fieldbook_sheet_hdfims}
\alias{add_fieldbook_sheet_hdfims}
\title{Add fieldbook sheet for HIDAP-AGROFIMS.}
\usage{
add_fieldbook_sheet_hdfims(file, fieldbook)
}
\arguments{
\item{file}{name of the file}

\item{fieldbook}{fieldbook data. A data frame.}
}
\description{
add a fieldbook sheet into a fieldbook file
}
