% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{add_installation_agrofims}
\alias{add_installation_agrofims}
\title{Create installation for hidap-agrofims}
\usage{
add_installation_agrofims(agronomic_crop_template, col_name = "Value",
  crop = "Potato", designFieldbook_agrofims = "",
  designFieldbook_agrofims_r = "", numPlantsPerPlot = "",
  numRowsPerPlot = "", numPlantsPerRow = "", plotSize = "",
  distancebwPlants = "", distanceBwRows = "", spaceBwPlants = "",
  spaceBwRows = "", planDensity = "", plotSpacing = "", rowSpacing = "",
  rowOrientation = "", spaceBwPlantsRow = "", hillSpacing = "",
  numsMsPlantPerPlot = "", fieldArea = "", expFieldMaxWidth = "",
  expFieldMaxLength = "", factor_hdafims1 = "", lvl_hdafims1 = "",
  factor_hdafims2 = "", lvl_hdafims2 = "", factor_hdafims3 = "",
  lvl_hdafims3 = "", factor_hdafims4 = "", lvl_hdafims4 = "",
  factor_hdafims5 = "", lvl_hdafims5 = "")
}
\description{
Create installation for hidap-agrofims
}

