% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{add_installation_sheet}
\alias{add_installation_sheet}
\title{Add values to Installation Sheet Data}
\usage{
add_installation_sheet(file = NA, crop_template = NA, col_name = "Value",
  exp_design = NA, genetic_design = NA, rep = NA, block = NA,
  exp_env = NA, plot_start_number = NA, n_plot_row = NA,
  n_plant_plot = NA, n_plant_row = NA, plot_size = NA,
  plant_density = NA, distance_plants = NA, distance_rows = NA,
  factor_name = NA, factor_name_1 = NA, factor_name_2 = NA,
  factor_name_3 = NA, factor_name_4 = NA, factor_name_5 = NA)
}
\arguments{
\item{file}{The file name of excel fieldbook}

\item{crop_template}{The crop template used to create Hidap Fieldbook}

\item{col_name}{The col name to paste values. By defaul is "Values"}

\item{exp_design}{The name of the excel fieldbook}

\item{genetic_design}{The genetic design}

\item{rep}{The number of repetition}

\item{block}{The number of block}

\item{exp_env}{The environment of the experiment}

\item{plot_start_number}{The plot start number}

\item{n_plot_row}{Number of plots per row}

\item{n_plant_plot}{The number of plant per plot}

\item{n_plant_row}{The number of plant per row}

\item{plot_size}{The number plant per row}

\item{plant_density}{The plant density}

\item{distance_plants}{The distance between plants}

\item{distance_rows}{The distance between rows}

\item{factor_name}{The name of factor}

\item{factor_name_1}{The first level of the factor}

\item{factor_name_2}{The second level of the factor}

\item{factor_name_3}{The third level of the factor}

\item{factor_name_4}{The fourth level of the factor}

\item{factor_name_5}{The fifth level of the factor}
}
\description{
This function allows to paste values in the Installation Excel (Fieldbook) Sheet
}

