% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{add_material_sheet}
\alias{add_material_sheet}
\title{Add values to Material_List Sheet Data}
\usage{
add_material_sheet(file = NA, crop_template = NA, crop, material_list)
}
\arguments{
\item{file}{The file name of excel fieldbook}

\item{crop_template}{The crop template used to create Hidap Fieldbook}

\item{crop}{The name of the crop}

\item{material_list}{The list of genotypes or materials}
}
\description{
This function allows to paste values into Material_List Excel (Fieldbook) Sheet
}

