% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{add_metadata_agrofims}
\alias{add_metadata_agrofims}
\title{Create metadata for hidap-agrofims}
\usage{
add_metadata_agrofims(agronomic_crop_template, col_name = "Value",
  experimentId = "", experimentName = "", experimentProjectName = "",
  startDate = "startDate", endDate = "endDate", Duration = "",
  typeExperiment = "", experimentObj = "", fundAgenType = "",
  fundName = "", contCenter = "", contCRP = "", contResearcher = "",
  fundLeadAgency = "", leadName = "", npersonnel = 1,
  personnel1Type = "", person1FirstName = "", person1LastName = "",
  person1Email = "", person1Afiliation = "", person1ORCID = "",
  personnel2Type = "", person2FirstName = "", person2LastName = "",
  person2Email = "", person2Afiliation = "", person2ORCID = "",
  personnel3Type = "", person3FirstName = "", person3LastName = "",
  person3Email = "", person3Afiliation = "", person3ORCID = "",
  personnel4Type = "", person4FirstName = "", person4LastName = "",
  person4Email = "", person4Afiliation = "", person4ORCID = "",
  personnel5Type = "", person5FirstName = "", person5LastName = "",
  person5Email = "", person5Afiliation = "", person5ORCID = "",
  sytpetype = "", syteName = "", siteID = "", countryName = "",
  villageName = "", nearestPopupPlace = "", inHighLevel = "",
  inSiteVegetation = "", inSiteDescNotes = "", croppingType = "",
  cropCommonNameMono = "", cropLatinNameMono = "",
  cropVarietyNameMono = "", cultivarNameMono = "", monoCropLocalName = "",
  numPreviousCrop = "", prevCropName = "", prevCropVar = "",
  subject = "", keywords = "", Embargo_date = "")
}
\description{
Create metadata for hidap-agrofims
}

