% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{add_metadata_sheet}
\alias{add_metadata_sheet}
\title{Add metadata to Weather and Soil Sheet Data}
\usage{
add_metadata_sheet(file = NA, crop_template = NA, soil_input, weather_input)
}
\arguments{
\item{file}{The file name of excel fieldbook}

\item{crop_template}{The crop template used to create Hidap Fieldbook}

\item{soil_input}{Logical. If soil input is selected (related to soil checkbox).}

\item{weather_input}{Logical. If weather input is selected (related to weather checkbox).}
}
\description{
This function allows to add weather and soil sheet data into Fieldbook Excel files.
}

