% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{add_minimal_sheet}
\alias{add_minimal_sheet}
\title{Add values to Minimal Sheet Data}
\usage{
add_minimal_sheet(file = NA, crop_template = NA, col_name = "Value",
  Trial_name = NA, crop = NA, type_trial = NA, begin_date = NA,
  end_date = NA, site_short_name = NA, country = NA)
}
\arguments{
\item{file}{The file name of excel fieldbook}

\item{crop_template}{The crop template used to create Hidap Fieldbook}

\item{col_name}{The col name to paste values. By defaul is "Values"}

\item{Trial_name}{The name of the excel fieldbook}

\item{crop}{The name of the crop}

\item{type_trial}{The type of trial}

\item{begin_date}{The begin date of the experiment}

\item{end_date}{The end date of the experiment}

\item{site_short_name}{The short name of the site}

\item{country}{The country name.}
}
\description{
This function allow to paste values in the Minimal Excel (Fieldbook) Sheet
}

