% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{add_varlist_sheet}
\alias{add_varlist_sheet}
\title{Add values to Var_List Sheet Data}
\usage{
add_varlist_sheet(file = NA, crop_template = NA, crop = NA,
  trait_list = NA)
}
\arguments{
\item{file}{The file name of excel fieldbook}

\item{crop_template}{The crop template used to create Hidap Fieldbook}

\item{crop}{The name of the crop}

\item{trait_list}{The input trait abbreviations using shinyTree}
}
\description{
This function allows to paste values into Crop_Management Excel (Fieldbook) Sheet
}

