% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{big_add_fieldbook_sheet}
\alias{big_add_fieldbook_sheet}
\title{Add values to Fieldbook Sheet Data}
\usage{
big_add_fieldbook_sheet(file, genotypes, big_sheet, fieldbook,
  big_crop_template)
}
\arguments{
\item{file}{The ID of fieldbook excel}

\item{genotypes}{A vector with genotypes to fill organoleptic forms}

\item{big_sheet}{The sheet used in Special or Big Trials}

\item{fieldbook}{A data.frame with the fieldbook data}
}
\description{
This function allows to paste values into Fieldbook Excel Sheet
}

