% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{biofer}
\alias{biofer}
\title{Create biofer for hidap-agrofims}
\usage{
biofer(biofertilizer_landLeveling_start_date,
  biofertilizer_landLeveling_end_date, biofertilizer_rhizobium_inoculum_strain,
  biofertilizer_quantity_inoculated, biofertilizer_inoculation_method,
  biofertilizer_product_formulation,
  biofertilizer_days_sowing_after_rhizobium_inocculation)
}
\description{
Create biofer for hidap-agrofims
}

