% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fbtrait_config.R
\name{create_fbtrait_config}
\alias{create_fbtrait_config}
\title{create_fbtrait_config}
\usage{
create_fbtrait_config(fieldbook, dictionary)
}
\arguments{
\item{fieldbook}{a fieldbook}

\item{dictionary}{a dictionary}
}
\value{
a dataframe
}
\description{
For Android fieldbook app
}

