% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subsamples.R
\name{create_subsamples}
\alias{create_subsamples}
\title{create_subsamples}
\usage{
create_subsamples(trait = "trait", timepoints = NULL, subsamples = NULL,
  subfun = "mean")
}
\arguments{
\item{trait}{a trait name}

\item{timepoints}{integer > 0, default is NULL}

\item{subsamples}{integer > 0, default is NULL}

\item{subfun}{one of "mean", "max", "min", "mode", "variance", "sum". Default is 'mean'.}
}
\value{
vector of trait labels
}
\description{
Creates a vector of labels for a trait name to help gather data for different time points (TP)
or sub samples (SS), that is plants from a plot or similar.
}
\details{
The helper variables are created according to a simple schema by suffixing TP plus sequential
number and SS plus sequential number separated by underscore (_). For subsamples an additional
variable is added that calculates a statistic to summarize the subsamples like 'mean'.
}
\author{
Reinhard Simon
}

