% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_fieldbook.R
\name{design_fieldbook}
\alias{design_fieldbook}
\title{Generic fieldbook design function}
\usage{
design_fieldbook(design = "(RCBD)", trt1 = letters[1:5], trt2 = NULL,
  r = 2, k = 2, trt1_label = "trt1", trt2_label = "trt2",
  factor_name = "FACTOR", factor_lvl1 = "level1", factor_lvl2 = "level2",
  factor_lvl3 = "level3", sub_design = "crd", number_row = NULL,
  number_col = NULL, set = NULL, male = NULL, female = NULL,
  maxRep = 20, series = 1, random = TRUE, first = TRUE, cont = FALSE,
  zigzag = FALSE, variables = NULL)
}
\arguments{
\item{design}{a statistical design}

\item{trt1}{vector treatment one}

\item{trt2}{vector second treatment; a controlled factor}

\item{r}{number of repetitions}

\item{k}{number of blocks}

\item{trt1_label}{string}

\item{trt2_label}{string}

\item{factor_name}{The name of the factor}

\item{factor_lvl1}{The first level of the factor}

\item{factor_lvl2}{The second level of the factor}

\item{factor_lvl3}{The third level of the factor}

\item{sub_design}{The sub design used by Split or Strip plot design.}

\item{number_row}{Wescott Design input. The number of rows.}

\item{number_col}{Wescott Design input. The number of columns.}

\item{set}{Number of sets (in case of genetic design)}

\item{male}{Males}

\item{female}{Females}

\item{maxRep}{maximum number of repetitions}

\item{series}{label series type}

\item{random}{to randomize or not}

\item{first}{to randomize or not the first repetition}

\item{cont}{continuouse labeling}

\item{zigzag}{order plot in serpentine}

\item{variables}{set of variables}
}
\value{
a dataframe
}
\description{
Generic fieldbook design function
}
