% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{experimental_design_label}
\alias{experimental_design_label}
\title{Experimental design Label}
\usage{
experimental_design_label(abbr_design = "RCBD")
}
\arguments{
\item{abbr_design}{The abbreviation of the experimental design}
}
\description{
Indicates which label correspond to statistical design abbreviation
}

