% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{genetic_design_label}
\alias{genetic_design_label}
\title{Genetic design Label}
\usage{
genetic_design_label(abbr_design = "NCI")
}
\arguments{
\item{abbr_design}{The abbreviation of the experimental design}
}
\description{
Indicates which label/full name correspond to every genetic design abbreviation
}

