% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_clean_date}
\alias{get_clean_date}
\title{Get clean dates from shiny inputs}
\usage{
get_clean_date(input_date)
}
\arguments{
\item{input_date}{input date}
}
\description{
Get clean dates from shiny inputs
}
