% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{harvest}
\alias{harvest}
\title{Create harvest for hidap-agrofims}
\usage{
harvest(harvest_start_date, harvest_end_date, crop_component_harvested,
  harvest_implement, harvest_make, harvest_model, harvest_animal_traction,
  harvest_humanPowered, harvest_motorized_traction)
}
\description{
Create harvest for hidap-agrofims
}

