% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{irrigation}
\alias{irrigation}
\title{Create irri for hidap-agrofims}
\usage{
irrigation(irrigationevent_start_date, irrigationevent_end_date,
  irrigation_system_type, irrigation_technique, surface_irrigation_technique,
  localized_irrigation_technique, irrigation_using_sprinkler_systems,
  irrigation_system_picture = "", irrigation_water_source,
  irrigation_water_source_distance, irrigation_bund_height,
  irrigation_percolation_rate, irrigation_equipment_depth,
  irrigation_well_depth, irrigation_area_covered_irrigation_system)
}
\description{
Create irri for hidap-agrofims
}

