% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_fieldbook.R
\name{is_control}
\alias{is_control}
\title{Add values to Fieldbook Sheet Data}
\usage{
is_control(material_list)
}
\arguments{
\item{material_list}{A data.frame with the material list data}
}
\description{
This function extract the control genotipes
}

