% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{land_des}
\alias{land_des}
\title{Create land des for hidap-agrofims}
\usage{
land_des(landLeveling_start_date, landLeveling_end_date, numPasses,
  operationsOrder, impl_type, animal_traction, humanPowered, motorized_traction,
  puddling_start_date, puddling_end_date, Penetrometer_in_field,
  puddling_depth_val, pud_animal_traction, pud_humanPowered,
  pud_motorized_traction, tillage_start_date, tillage_end_date, till_technique,
  till_depth_method, till_depth, till_total_op_season, till_impl_type,
  till_animal_traction, till_humanPowered, till_motorized_traction,
  liming_start_date, liming_end_date, lim_material, lim_quantity,
  lim_description)
}
\description{
Create land des for hidap-agrofims
}

