% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{material_list_parental}
\alias{material_list_parental}
\title{material_list_parental template}
\format{A list of installation sheet data in HiDAP offline}
\source{
International Potato Center CIP
}
\usage{
material_list_parental
}
\description{
Template for parental list
}
\references{
This data is related to HIDAP desktop in order to assist genetic studiess
}
\keyword{genetic}
\keyword{parental}
\keyword{potato}
