% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{mulch}
\alias{mulch}
\title{Create mulching for hidap-agrofims}
\usage{
mulch(mulch_start_date, mulch_end_date, mulch_type, mulch_thickness,
  mulch_amountPerSq, mulch_color, mulch_percCoverage, mulch_remove_start_date,
  mulch_remove_end_date, mulch_make, mulch_model, mulch_animal_traction,
  mulch_humanPowered, mulch_motorized_traction, residue_cropType,
  residue_technique, residue_incorp_depth, residue_aboveGroundMoisture,
  residue_aboveGroundAmount)
}
\description{
Create mulching for hidap-agrofims
}

