% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection_tools.R
\name{organoleptic_form}
\alias{organoleptic_form}
\title{Organoleptic forms to collect sensorial data in participatory varietal selection of potato and sweetpotato trials.}
\usage{
organoleptic_form(form_template, genotypes, geno_alias, n_forms = 1)
}
\arguments{
\item{form_template}{Data frame containing the form template to collect sensorial data.}

\item{genotypes}{The genotypes.}

\item{geno_alias}{The alias of genotypes.}

\item{n_forms}{Number of forms. By default is 1.}
}
\value{
A data frame with the organoleptic form
}
\description{
On farm methodologies requiere short and large organoleptic screenings where agricultors give sensorial scores
(taste, apparearence and texture) #' for selecting best clones or genotypes. In order to achive, this function
}
