% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{pestdis}
\alias{pestdis}
\title{Create pest and disease for hidap-agrofims}
\usage{
pestdis(disease_observation_date, disease_name, disease_plant_parts_affected,
  disease_percentage_experiement_affected, disease_damages_notes, disease_notes,
  pest_type, pest_name, pest_damage_notes, pest_notes, pestcontrol_start_date,
  pestcontrol_end_date, pest_control_technique, pesticide_application_depth,
  pesticide_amount, pest_image = "", pest_control_applications_totnumber,
  pest_control_details, chemical_pest_control_equipment,
  pesticide_implement_make, pesticide_implement_model,
  pesticide_animal_traction, pesticide_humanPowered,
  pesticide_motorized_traction)
}
\description{
Create pest and disease for hidap-agrofims
}

