% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{plant}
\alias{plant}
\title{Create planting for hidap-agrofims}
\usage{
plant(planting_start_date, planting_end_date, planting_directSeeding,
  planting_seedingTech, planting_ageSeeding, planting_manual,
  planting_animal_traction, planting_motorized_traction, planting_rowDistance,
  planting_seedingRate, planting_seedPerhill, planting_distance,
  planting_distribution)
}
\description{
Create planting for hidap-agrofims
}

