% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.design.R
\name{randomize.design}
\alias{randomize.design}
\title{Design a fieldbook.}
\usage{
randomize.design(design = "(CRD)", matl, reps, msite = FALSE, lbls,
  checkl = NULL, bsize = 2, adfn = NULL, adfl = NULL, startn = 1,
  seed = 0, randM = "Super-duper")
}
\description{
param design statistical design
param matl list of germplasm
param reps number of plot repetitions
param msite logical, is this a mother site in a mother baby trial
param lbls labels for germplasm
param checkl list of check germplasm
param bsize block size
param adfn name of additional factor
param adfl levels of additional factor
param startn start number
param seed random seed
param randM randomization method
importFrom stringr str_detect
importFrom agricolae design.crd design.rcbd design.lsd
author Raul Eyzaguirre, Reinhard Simon
return data.frame
export
}

