% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fbdesgin_mtl_files}
\alias{fbdesgin_mtl_files}
\title{List of the dbf files}
\usage{
fbdesgin_mtl_files()
}
\description{
List of the dbf files
}
\section{Functions}{
\itemize{
\item \code{fbdesgin_mtl_files}: the table from M.List DB an write into the computer
}}

