% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_vars_subsamples.R
\name{replace_vars_subsamples}
\alias{replace_vars_subsamples}
\title{replace_vars_subsamples}
\usage{
replace_vars_subsamples(traits, config_subsamples)
}
\arguments{
\item{traits}{a vector of traits}

\item{config_subsamples}{a dataframe}
}
\value{
a vector of trait labels
}
\description{
replace_vars_subsamples
}

