% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_design.R
\name{server_design}
\alias{server_design}
\title{server_design}
\usage{
server_design(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Design field book
}
\author{
Omar Benites
}

