% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_design_big.R
\name{server_design_big}
\alias{server_design_big}
\title{server_design_big}
\usage{
server_design_big(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server_design_big provides a server functionalities for Special or Big modules like PVS and ABS
}
\details{
Design a fieldbook
}
\author{
Omar Benites
}

