% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_design_big.R
\name{ui_design_big}
\alias{ui_design_big}
\title{shiny UI element big}
\usage{
ui_design_big(type = "tab", title = "Special Modules",
  name = "phenotype_fieldbook_design")
}
\arguments{
\item{type}{of ui Element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}
}
\description{
User interface function to build Special Modules like PVS or ABS in HIDAP.
}
\details{
returns a re-usable user interface element big
}
\author{
Reinhard Simon
}
