% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fieldbook_params.R
\name{ui_fieldbook_params}
\alias{ui_fieldbook_params}
\title{UI material list paramters}
\usage{
ui_fieldbook_params(name = "phenotype_fieldbook_design")
}
\arguments{
\item{name}{character}
}
\description{
An interface to parameters
}
\author{
Reinhard Simon
}

