% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_xlsx_agrofims.R
\name{add_metadata_agrofims}
\alias{add_metadata_agrofims}
\title{Create metadata for hidap-agrofims}
\usage{
add_metadata_agrofims(agronomic_crop_template, col_name = "Value",
  experimentId = "", experimentName = "", experimentProjectName = "",
  startDate = "startDate", endDate = "endDate", Duration = "",
  designFieldbook_typeExperiment = "", experimentObj = "", fundName = "",
  personnel1Type = "", person1FirstName = "",
  designFieldbook_fundLeadAgency = "", leadName = "",
  person1LastName = "", person1Email = "", person1Afiliation = "",
  person1ORCID = "", instAbreviation = "", contCenter = "",
  contCRP = "", contResearcher = "", sytpetype = "", syteName = "",
  siteID = "", countryName = "", admin1 = "", admin2 = "",
  villageName = "", elevation = "", latitude = "", longitude = "",
  descNotes = "", croppingType, cropLatinNameMono = "",
  cropCommonNameMono = "", cultivarNameMono = "",
  cropVarietyNameMono = "", subject = "", keywords = "",
  Embargo_date = "")
}
\description{
Create metadata for hidap-agrofims
}

