% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_fieldbook.R
\name{design_fieldbook}
\alias{design_fieldbook}
\title{Generic fieldbook design function}
\usage{
design_fieldbook(design, trt1, trt2 = NULL, r = 2, k, series = 2,
  random = TRUE, first = FALSE, cont = FALSE, variables = NULL)
}
\arguments{
\item{design}{a statistical design}

\item{trt1}{label of treatment one; mostly 'GENOTYPE'}

\item{trt2}{label of second treatment; a controlled factor}

\item{r}{number of repetitions}

\item{k}{number of blocks}

\item{series}{label series type}

\item{random}{to randomize or not}

\item{first}{to randomize or not the first repetition}

\item{cont}{continuouse labeling}

\item{variables}{set of variables}
}
\value{
a dataframe
}
\description{
Generic fieldbook design function
}

