% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_fieldbook.R
\name{design_fieldbook}
\alias{design_fieldbook}
\title{design_fieldbook}
\usage{
design_fieldbook(design = "(RCBD)", trt1 = letters[1:5], trt2 = NULL,
  r = 2, k = 2, trt1_label = "trt1", trt2_label = "trt2", maxRep = 20,
  series = 2, random = TRUE, first = FALSE, cont = FALSE,
  zigzag = FALSE, variables = NULL)
}
\arguments{
\item{design}{a statistical design}

\item{trt1}{vector treatment one}

\item{trt2}{vector second treatment; a controlled factor}

\item{r}{number of repetitions}

\item{k}{number of blocks}

\item{trt1_label}{string}

\item{trt2_label}{string}

\item{maxRep}{maximum number of repetitions}

\item{series}{label series type}

\item{random}{to randomize or not}

\item{first}{to randomize or not the first repetition}

\item{cont}{continuouse labeling}

\item{zigzag}{order plot in serpentine}

\item{variables}{set of variables}
}
\value{
a dataframe
}
\description{
Generic fieldbook design function
}
\author{
Reinhard Simon
}

