% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.design.R
\name{randomize.design}
\alias{randomize.design}
\title{Design a fieldbook.}
\usage{
randomize.design(design = "(CRD)", matl, reps, msite = FALSE, lbls,
  checkl = NULL, bsize = 2, adfn = NULL, adfl = NULL, startn = 1,
  seed = 0, randM = "Super-duper")
}
\arguments{
\item{design}{statistical design}

\item{matl}{list of germplasm}

\item{reps}{number of plot repetitions}

\item{msite}{logical, is this a mother site in a mother baby trial}

\item{lbls}{labels for germplasm}

\item{checkl}{list of check germplasm}

\item{bsize}{block size}

\item{adfn}{name of additional factor}

\item{adfl}{levels of additional factor}

\item{startn}{start number}

\item{seed}{random seed}

\item{randM}{randomization method}
}
\value{
data.frame
}
\description{
Design a fieldbook.
}
\author{
Raul Eyzaguirre, Reinhard Simon
}

