% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_design.R
\name{server_design}
\alias{server_design}
\title{server_design}
\usage{
server_design(input, output, session, dom = "hot_fieldbook_design", values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{dom}{target dom element name}

\item{values}{reactive values}
}
\description{
Design a fieldbook
}
\author{
Reinhard Simon
}

