% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_module.R
\name{ui_fieldbook}
\alias{ui_fieldbook}
\title{shiny UI element}
\usage{
ui_fieldbook(type = "tab", title = "Design field book",
  name = "phenotype_fieldbook_design", output = "hot_fieldbook_design")
}
\arguments{
\item{type}{of ui Element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}

\item{output}{name of output element
export}
}
\description{
returns a re-usable user interface element
}
\author{
Reinhard Simon
}

